ref: make settings behave using python-decouple instead of pydantic
this is a huge refactor. it involves dropping BaseSettings usage from pydantic and relying on a more simple approach: load from environment if available, load a sensible default if missing, hardcode testing and development variables.

This adds some tests to ensure the proper database is being used. It also refactors the CLI to reflect new usage